# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface OrderItem {
  productId: number;
  name: string;
  /** @minimum 1 */
  quantity: number;
  size: string;
  milk: string;
  unitPrice: number;
}

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  pickup: 'pickup',
  'dine-in': 'dine-in',
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  orderType: OrderInputOrderType;
  /** @minItems 1 */
  items: OrderItem[];
  /** @minimum 0 */
  total: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  preparing: 'preparing',
  ready: 'ready',
} as const;

export interface Order {
  id: number;
  status: OrderStatus;
  prepMinutes: number;
  eta: string;
  total: number;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  name: string;
  email: string;
  date: string;
  /** @minLength 1 */
  time: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  partySize: number;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
} as const;

export interface Booking {
  id: number;
  name: string;
  date: string;
  time: string;
  partySize: number;
  status: BookingStatus;
  createdAt: string;
}