# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 1 */
  name: string;
  email: string;
  date: Date;
  /** @minLength 1 */
  time: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  partySize: number;
}