# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  prepMinutes: number;
  eta: Date;
  total: number;
  createdAt: Date;
}