# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from './orderInputOrderType';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  orderType: OrderInputOrderType;
  /** @minItems 1 */
  items: OrderItem[];
  /** @minimum 0 */
  total: number;
}