# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  name: string;
  /** @minimum 1 */
  quantity: number;
  size: string;
  milk: string;
  unitPrice: number;
}